+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/view/MiniAppView.kt
@@ -63,4 +63,12 @@ abstract class MiniAppView internal constructor() {
         fromCache: Boolean = false,
         onComplete: (MiniAppDisplay?, MiniAppSdkException?) -> Unit
     )
+
+    /**
+     * Send a generic message to MiniApp using
+     * [com.rakuten.tech.mobile.miniapp.js.NativeEventType.MINIAPP_RECEIVE_JSON_INFO].
+     * @param message the content that will send to the MiniApp
+     * @param onFailed invoked when the validation does not pass to send the message
+     */
+    abstract fun sendJsonToMiniApp(message: String, onFailed: () -> Unit)
 }