+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/js/MessageBridgeRatDispatcherSpec.kt
@@ -37,8 +37,13 @@ class MessageBridgeRatDispatcherSpec {
         ActionType.FILE_DOWNLOAD.action to Actype.FILE_DOWNLOAD,
         ActionType.GET_UNIQUE_ID.action to Actype.GET_UNIQUE_ID,
         ActionType.GET_MESSAGING_UNIQUE_ID.action to Actype.GET_MESSAGING_UNIQUE_ID,
-        ActionType.GET_MAUID.action to Actype.GET_MAUID
-
+        ActionType.GET_MAUID.action to Actype.GET_MAUID,
+        ActionType.SECURE_STORAGE_SET_ITEMS.action to Actype.SECURE_STORAGE_SET_ITEMS,
+        ActionType.SECURE_STORAGE_GET_ITEM.action to Actype.SECURE_STORAGE_GET_ITEM,
+        ActionType.SECURE_STORAGE_REMOVE_ITEMS.action to Actype.SECURE_STORAGE_REMOVE_ITEMS,
+        ActionType.SECURE_STORAGE_CLEAR.action to Actype.SECURE_STORAGE_CLEAR,
+        ActionType.SECURE_STORAGE_SIZE.action to Actype.SECURE_STORAGE_SIZE,
+        ActionType.JSON_INFO.action to Actype.JSON_INFO,
     )
 
     @Test