+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/view/MiniAppViewSpec.kt
@@ -89,7 +89,7 @@ open class MiniAppViewSpec {
             mock(),
             null,
             null,
-            ""
+            "",
         )
 
         miniAppConfig.shouldBeInstanceOf<MiniAppConfig>()