+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/miniapptabs/DemoAppMainActivity.kt
@@ -10,12 +10,15 @@ import androidx.navigation.NavController
 import androidx.navigation.ui.setupActionBarWithNavController
 import com.rakuten.tech.mobile.miniapp.testapp.R
 import com.rakuten.tech.mobile.miniapp.testapp.databinding.MiniAppMainLayoutBinding
+import com.rakuten.tech.mobile.miniapp.view.MiniAppView
 import com.rakuten.tech.mobile.testapp.ui.base.BaseActivity
 import com.rakuten.tech.mobile.testapp.ui.miniapptabs.extensions.setupWithNavController
 import com.rakuten.tech.mobile.testapp.ui.miniapptabs.fragments.MiniAppDisplayFragment
 import com.rakuten.tech.mobile.testapp.ui.settings.AppSettings
 import kotlinx.android.synthetic.main.mini_app_main_layout.*
 
+
+val miniAppIdAndViewMap = hashMapOf<Pair<Int, String>, MiniAppView>()
 
 class DemoAppMainActivity : BaseActivity() {
     private var currentNavController: LiveData<NavController>? = null
@@ -46,6 +49,7 @@ class DemoAppMainActivity : BaseActivity() {
             intent = intent
         )
 
+
         controller.observe(this) { navController ->
             setupActionBarWithNavController(navController)
 
@@ -63,8 +67,11 @@ class DemoAppMainActivity : BaseActivity() {
 
     private val onDestinationChangedListener =
         NavController.OnDestinationChangedListener { controller, destination, arguments ->
-            Log.d("TAG", "controller: $controller, destination: $destination, arguments: $arguments")
-            Log.d("TAG","controller graph: ${controller.graph.id}")
+            Log.d(
+                "TAG",
+                "controller: $controller, destination: $destination, arguments: $arguments"
+            )
+            Log.d("TAG", "controller graph: ${controller.graph.id}")
 
             //Set the mini app settings depending on the tab.
             when (controller.graph.id) {
@@ -101,6 +108,8 @@ class DemoAppMainActivity : BaseActivity() {
         bottomNavigationView.selectedItemId = page
     }
 
+    fun getCurrentSelectedId() = bottomNavigationView.selectedItemId
+
     companion object {
         private const val PAGE_1 = R.id.nav_tab_0
         private const val PAGE_2 = R.id.nav_tab_1
@@ -110,6 +119,7 @@ class DemoAppMainActivity : BaseActivity() {
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         super.onActivityResult(requestCode, resultCode, data)
+
         getCurrentVisibleFragment()?.let { fragment ->
             if (fragment is MiniAppDisplayFragment) {
                 fragment.handleOnActivityResult(requestCode, resultCode, data)
@@ -135,7 +145,6 @@ class DemoAppMainActivity : BaseActivity() {
             if (fragment is MiniAppDisplayFragment) {
                 fragment.onBackPressed()
             } else super.onBackPressed()
-
         }
     }