+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/settings/AppSettings.kt
@@ -210,7 +210,7 @@ class AppSettings private constructor(context: Context) {
         cache.rasConfigData.saveTab2Data()
     }
 
-    fun saveCurrentAppInfoList(miniAppInfoList: List<MiniAppInfo>){
+    fun saveCurrentAppInfoList(miniAppInfoList: List<MiniAppInfo>) {
         cache.rasConfigData.saveCurrentMiniAppInfoList(miniAppInfoList, miniAppInfoListKey)
     }
 
@@ -269,7 +269,7 @@ class AppSettings private constructor(context: Context) {
         cache.rasConfigData.clearTempData()
     }
 
-    fun clearAllMiniAppInfoList(){
+    fun clearAllMiniAppInfoList() {
         cache.rasConfigData.clearAllMiniAppInfoList()
     }